[RUNTIME] Enable auto conversion String->DLDataType
